+++ lib/savon/client.rb
@@ -82,7 +82,7 @@ module Savon
       soap.wsse = wsse
 
       response = SOAP::Request.new(config, http, soap).response
-      set_cookie response.http.headers
+      http.set_cookies(response.http)
 
       if wsse.verify_response
         WSSE::VerifySignature.new(response.http.body).verify!
@@ -98,23 +98,6 @@ module Savon
 
     # Accessor for the original self of a given block.
     attr_accessor :original_self
-
-    # Passes a cookie from the last request +headers+ to the next one.
-    def set_cookie(headers)
-      if headers["Set-Cookie"]
-        @cookies ||= {}
-        #handle single or multiple Set-Cookie Headers as returned by Rack::Utils::HeaderHash in HTTPI
-        set_cookies = [headers["Set-Cookie"]].flatten
-        set_cookies.each do |set_cookie|
-          # use the cookie name as the key to the hash to allow for cookie updates and seperation
-          # set the value to name=value (for easy joining), stopping when we hit the Cookie options
-          @cookies[set_cookie.split('=').first] = set_cookie.split(';').first
-        end
-
-        http.headers["Cookie"] = @cookies.values.join(';')
-      end
-
-    end
 
     # Expects an Array of +args+ and returns an Array containing the namespace (might be +nil+),
     # the SOAP input and a Hash of attributes for the input tag (which might be empty).